Move `GuiMakeSpendInfo` into the types file
