+++ src/components/themed/AddressTile.js
@@ -10,8 +10,8 @@ import FontAwesome5 from 'react-native-vector-icons/FontAwesome5'
 import { CURRENCY_PLUGIN_NAMES } from '../../constants/WalletAndCurrencyConstants'
 import s from '../../locales/strings.js'
 import { checkExpiredFioAddress, checkPubAddress } from '../../modules/FioAddress/util'
-import { type GuiMakeSpendInfo } from '../../reducers/scenes/SendConfirmationReducer.js'
 import { connect } from '../../types/reactRedux.js'
+import { type GuiMakeSpendInfo } from '../../types/types.js'
 import { AddressModal } from '../modals/AddressModal'
 import { paymentProtocolUriReceived } from '../modals/paymentProtocolUriReceived.js'
 import { ScanModal } from '../modals/ScanModal.js'
