+++ src/modules/UI/scenes/Plugins/EdgeProvider.js
@@ -17,10 +17,9 @@ import { type WalletListResult, WalletListModal } from '../../../../components/m
 import { Airship, showError, showToast } from '../../../../components/services/AirshipInstance.js'
 import { SEND } from '../../../../constants/SceneKeys.js'
 import s from '../../../../locales/strings'
-import type { GuiMakeSpendInfo } from '../../../../reducers/scenes/SendConfirmationReducer.js'
 import { type GuiPlugin, type GuiPluginQuery } from '../../../../types/GuiPluginTypes.js'
 import { type Dispatch, type RootState } from '../../../../types/reduxTypes.js'
-import { type GuiWallet } from '../../../../types/types.js'
+import { type GuiMakeSpendInfo, type GuiWallet } from '../../../../types/types.js'
 
 type EdgeReceiveAddress = {
   publicAddress?: string,
